Version 1.0
Breaking change: APLImageCache used to scaleToFill its contents. This was later changed to be ApectFit or AspectFill depending on the image cache, which broke existing users. This change exposes a setting [APLImageCache setCacheContentMode:] where you can configure the content mode to be used. Since scaleToFill is UIViewContentMode's default value, this may or may not be a breaking change for you.